requirements.txt
requests
redis
pymongo
cassandra-driver
neo4j
